repair: Move row_repair hashing logic to separate class and file.
